requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
validators==0.22.0
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0